Cleaned up skip functions from e2e framework
This patch moves skip-related functions from framework/util.go to
another file.
